Refactor default asset storage stuff
